# server/docs/swagger.json
{
    "consumes": [
        "multipart/form-data"
    ],
    "produces": [
        "json image/gif"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A golang server to create gifs from images.",
        "title": "Makesweet golang server",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/gif/billboard": {
            "post": {
                "description": "Use image from form to make a approaching billboard gif",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "json image/gif"
                ],
                "tags": [
                    "Gif"
                ],
                "summary": "Create a billboard gif",
                "parameters": [
                    {
                        "type": "file",
                        "description": "A png or jpg image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "fast",
                            "slow"
                        ],
                        "type": "string",
                        "default": "slow",
                        "description": "Gif mode, can be 'fast' or 'slow', default is 'slow'",
                        "name": "mode",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated Gif",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Fail to load image from form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Fail to generate gif",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/gif/circuit": {
            "post": {
                "description": "Use image from form to make a zooming circuit board gif",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "json image/gif"
                ],
                "tags": [
                    "Gif"
                ],
                "summary": "Create a circuit board gif",
                "parameters": [
                    {
                        "type": "file",
                        "description": "A png or jpg image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "fast",
                            "slow"
                        ],
                        "type": "string",
                        "default": "slow",
                        "description": "Gif mode, can be 'fast' or 'slow', default is 'slow'",
                        "name": "mode",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated Gif",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Fail to load image from form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Fail to generate gif",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/gif/custom": {
            "post": {
                "description": "Use images and a template from form to make a custom gif",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "json image/gif"
                ],
                "tags": [
                    "Gif"
                ],
                "summary": "Create a custom gif",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "csv",
                        "description": "A png or jpg image array",
                        "name": "images",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "A zip template file",
                        "name": "template",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "fast",
                            "slow"
                        ],
                        "type": "string",
                        "default": "slow",
                        "description": "Gif mode, can be 'fast' or 'slow', default is 'slow'",
                        "name": "mode",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated Gif",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Fail to load images from form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Fail to generate gif",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/gif/flag": {
            "post": {
                "description": "Use image from form to make a waving flag gif",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "json image/gif"
                ],
                "tags": [
                    "Gif"
                ],
                "summary": "Create a flag gif",
                "parameters": [
                    {
                        "type": "file",
                        "description": "A png or jpg image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "fast",
                            "slow"
                        ],
                        "type": "string",
                        "default": "slow",
                        "description": "Gif mode, can be 'fast' or 'slow', default is 'slow'",
                        "name": "mode",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated Gif",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Fail to load image from formData",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Fail to generate gif",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/gif/flying-bear": {
            "post": {
                "description": "Use image from form to make a flying bear gif",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "json image/gif"
                ],
                "tags": [
                    "Gif"
                ],
                "summary": "Create a flying bear gif",
                "parameters": [
                    {
                        "type": "file",
                        "description": "A png or jpg image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "fast",
                            "slow"
                        ],
                        "type": "string",
                        "default": "slow",
                        "description": "Gif mode, can be 'fast' or 'slow', default is 'slow'",
                        "name": "mode",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated Gif",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Fail to load image from form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Fail to generate gif",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/gif/heart-locket": {
            "post": {
                "description": "Use image-lef and image-right files from form to make a opening heart locket gif",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "json image/gif"
                ],
                "tags": [
                    "Gif"
                ],
                "summary": "Create a heart locket gif",
                "parameters": [
                    {
                        "type": "file",
                        "description": "A png or jpg image to left half",
                        "name": "image-left",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "A png or jpg image to right half",
                        "name": "image-right",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "standard",
                            "christmas"
                        ],
                        "type": "string",
                        "default": "standard",
                        "description": "Variant of heart locket gif",
                        "name": "variant",
                        "in": "formData"
                    },
                    {
                        "enum": [
                            "fast",
                            "slow"
                        ],
                        "type": "string",
                        "default": "slow",
                        "description": "Gif mode, can be 'fast' or 'slow', default is 'slow'",
                        "name": "mode",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated Gif",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Fail to load image from form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Fail to generate gif",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/gif/nesting-doll": {
            "post": {
                "description": "Use three images to create a nesting doll gif",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "json image/gif"
                ],
                "tags": [
                    "Gif"
                ],
                "summary": "Create a nesting doll gif",
                "parameters": [
                    {
                        "type": "file",
                        "description": "A png or jpg image to the left doll",
                        "name": "image-left",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "A png or jpg image to the mid doll",
                        "name": "image-mid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "A png or jpg image to the right doll",
                        "name": "image-right",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "fast",
                            "slow"
                        ],
                        "type": "string",
                        "default": "slow",
                        "description": "Gif mode, can be 'fast' or 'slow', default is 'slow'",
                        "name": "mode",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated Gif",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Fail to load images from formData",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Fail to generate gif",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}